WIP trying and failing to make progress on dv writes
